Made Rule class as abstract and project structure refactoring
