+++ src/main/java/agh/edu/hermes/services/SlaService.java
@@ -3,7 +3,7 @@ package agh.edu.hermes.services;
 import agh.edu.hermes.checker.SlaViolationChecker;
 import agh.edu.hermes.storages.RuleStorage;
 import agh.edu.hermes.storages.Sla;
-import agh.edu.hermes.types.Rule;
+import agh.edu.hermes.types.base.Rule;
 import agh.edu.hermes.types.SlaRule;
 import org.springframework.stereotype.Service;
