+++ src/main/java/agh/edu/hermes/storages/Policies.java
@@ -1,7 +1,7 @@
 package agh.edu.hermes.storages;
 
 import agh.edu.hermes.types.PolicyRule;
-import agh.edu.hermes.types.Rule;
+import agh.edu.hermes.types.base.Rule;
 
 import java.util.ArrayList;
 import java.util.List;
