+++ src/main/java/agh/edu/hermes/storages/Sla.java
@@ -1,7 +1,7 @@
 package agh.edu.hermes.storages;
 
 
-import agh.edu.hermes.types.Rule;
+import agh.edu.hermes.types.base.Rule;
 import agh.edu.hermes.types.SlaRule;
 
 import java.util.ArrayList;
