+++ src/main/java/agh/edu/hermes/types/attributes/RuleSubject.java
@@ -1,4 +1,4 @@
-package agh.edu.hermes.types;
+package agh.edu.hermes.types.attributes;
 
 public enum RuleSubject {
     RAM,
